Patched normalizeWidth. This is used in the Charter package.
The old implementation was like this:
tt := min + ((max - min) * ( (t - minValue) / (maxValue - minValue))) asInteger.

The problem with this is that the normalization will return for the lowest value, the specified min value.
So, if we had two numbers 10 and 5 to be normalized between 0 and 100, we would get:
- 10 => 100
- 5 => 0

This rule works for colors, but for numbers, we want that 5 to be 50.

So, the current implementation looks like:
tt := t * (max-min) / maxValue.