ExploraBuilder: changed the layout application on every elements in the view... We can collect the elements in that builder with the nodes collection.
RTMetricNormalizer: change in getNormalizedValues, when the minValue = maxValue, (when only one element for example)  collect the value of the min instead of the evaluation of the model by the block. In that case we always stay in the bound of min and max when the minValue = maxValue.

Fixed normalizeWidth:min:max:using to call the getNormalized method like it is working for height, size or color already 

Former-commit-id: fe3bfaf78d0cf37625fc6100ee9bc41dc4479d27