Yuriy found this:
On Jun 9, 2015, at 3:47 PM, Yuriy Tymchuk <yuriy.tymchuk@me.com> wrote:

Hi everyone,

I guess because of a rapid evolution some parts of Roassal are weird. For example RTAbstractLine>>#updateFor:trachelShape: implements the shape drawing, so theoretically RTLine has to only override #trachelShapeClass to return TRLine. Instead RTLine overrides #trachelShapeFor: and does it incorrectly, because width cannot be used as a block but only as a number.

We have to handle this somehow, because 1) RTLine works incorrectly. 2) RTLine ignores the implementation strategy coded into RTAbstractLine.

Uko

This is now fixed!